+++ android/democmds/jni/democmds_java_wrap.cpp
@@ -769,13 +769,13 @@ SWIGEXPORT jint JNICALL Java_democmds_core_democmdsJNI_MgCoreView_1getShapeCount
 SWIGEXPORT jint JNICALL Java_democmds_core_democmdsJNI_MgCoreView_1getChangeCount(JNIEnv *jenv, jclass jcls, jlong jarg1, jobject jarg1_) {
   jint jresult = 0 ;
   MgCoreView *arg1 = (MgCoreView *) 0 ;
-  int result;
+  long result;
   
   (void)jenv;
   (void)jcls;
   (void)jarg1_;
   arg1 = *(MgCoreView **)&jarg1; 
-  result = (int)(arg1)->getChangeCount();
+  result = (long)(arg1)->getChangeCount();
   jresult = (jint)result; 
   return jresult;
 }
@@ -784,13 +784,13 @@ SWIGEXPORT jint JNICALL Java_democmds_core_democmdsJNI_MgCoreView_1getChangeCoun
 SWIGEXPORT jint JNICALL Java_democmds_core_democmdsJNI_MgCoreView_1getDrawCount(JNIEnv *jenv, jclass jcls, jlong jarg1, jobject jarg1_) {
   jint jresult = 0 ;
   MgCoreView *arg1 = (MgCoreView *) 0 ;
-  int result;
+  long result;
   
   (void)jenv;
   (void)jcls;
   (void)jarg1_;
   arg1 = *(MgCoreView **)&jarg1; 
-  result = (int)((MgCoreView const *)arg1)->getDrawCount();
+  result = (long)((MgCoreView const *)arg1)->getDrawCount();
   jresult = (jint)result; 
   return jresult;
 }
