+++ ios/output/TouchVG/mgcoreview.h
@@ -40,10 +40,10 @@ struct MgCoreView
     virtual int getShapeCount() = 0;
 
     //! 返回静态图形改变次数，可用于检查是否需要保存
-    virtual int getChangeCount() = 0;
+    virtual long getChangeCount() = 0;
     
     //! 返回已绘制次数，可用于录屏
-    virtual int getDrawCount() const = 0;
+    virtual long getDrawCount() const = 0;
 
     //! 返回选中的图形个数
     virtual int getSelectedShapeCount() = 0;
